+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotContextImpl.java
@@ -102,7 +102,6 @@ import com.oracle.truffle.api.frame.FrameInstance;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.interop.UnsupportedMessageException;
-import com.oracle.truffle.api.nodes.EncapsulatingNodeReference;
 import com.oracle.truffle.api.nodes.LanguageInfo;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.source.Source;
@@ -699,7 +698,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
 
     public synchronized void explicitEnter() {
         try {
-            PolyglotContextImpl prev = engine.enter(this, true, engine.getUncachedLocation(), true, false);
+            PolyglotContextImpl prev = engine.enter(this);
             PolyglotThreadInfo current = getCachedThreadInfo();
             assert current.getThread() == Thread.currentThread();
             current.explicitContextStack.addLast(prev);
@@ -722,7 +721,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
             if (stack.isEmpty() || current.getThread() == null) {
                 throw PolyglotEngineException.illegalState("The context is not entered explicity. A context can only be left if it was previously entered.");
             }
-            engine.leave(stack.removeLast(), this, true);
+            engine.leave(stack.removeLast(), this);
         } catch (Throwable t) {
             throw PolyglotImpl.guestToHostException(engine, t);
         }
@@ -744,8 +743,12 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
         }
     }
 
+    /**
+     * Use to enter context if it's guaranteed to be called rarely and configuration flexibility is
+     * needed. Otherwise use {@link PolyglotEngineImpl#enter(PolyglotContextImpl)}.
+     */
     @TruffleBoundary
-    PolyglotContextImpl enterThreadChanged(boolean notifyEnter, Node safepointLocation, boolean enterReverted, boolean pollSafepoint, boolean deactivateSafepoints) {
+    PolyglotContextImpl enterThreadChanged(boolean notifyEnter, boolean enterReverted, boolean pollSafepoint, boolean deactivateSafepoints) {
         PolyglotThreadInfo enteredThread = null;
         PolyglotContextImpl prev = null;
         try {
@@ -872,14 +875,14 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
              */
             if (pollSafepoint) {
                 try {
-                    TruffleSafepoint.pollHere(safepointLocation);
+                    TruffleSafepoint.pollHere(engine.getUncachedLocation());
                 } catch (Throwable t) {
                     /*
                      * Just in case a safepoint makes the enter fail we need to leave the context
                      * again.
                      */
                     if (enteredThread != null) {
-                        engine.leave(prev, this, notifyEnter);
+                        this.leaveThreadChanged(prev, notifyEnter, true);
                     }
                     throw t;
                 }
@@ -934,6 +937,11 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
         }
     }
 
+    /**
+     * Use to leave a context if its guaranteed to be called rarely and configuration flexibility is
+     * needed. Otherwise use
+     * {@link PolyglotEngineImpl#leave(PolyglotContextImpl, PolyglotContextImpl)}.
+     */
     @TruffleBoundary
     PolyglotThreadInfo leaveThreadChanged(PolyglotContextImpl prev, boolean notifyLeft, boolean entered) {
         PolyglotThreadInfo info;
@@ -1161,40 +1169,45 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
         }
     }
 
-    Object migrateValue(Node parentNode, Object value, PolyglotContextImpl valueContext) {
-        // migration of guest primitives is already handled.
-        if (PolyglotImpl.isGuestPrimitive(value)) {
-            // allowed to be passed freely
-            return value;
-        } else if (engine.host.isHostObject(value)) {
-            return engine.host.migrateHostObject(getHostContextImpl(), value);
-        } else if (engine.host.isHostProxy(value)) {
-            return engine.host.migrateHostProxy(getHostContextImpl(), value);
-        } else if (valueContext == null) {
-            /*
-             * The only way this can happen is with Value.asValue(TruffleObject). If it happens
-             * otherwise, its wrong.
-             */
-            assert value instanceof TruffleObject;
-            return value;
-        } else {
-            CompilerDirectives.transferToInterpreterAndInvalidate();
-            Node nodeContext = parentNode;
-            if (nodeContext == null || nodeContext.getRootNode() == null) {
-                nodeContext = EncapsulatingNodeReference.getCurrent().get();
-            }
-            throw new ValueMigrationException(String.format("The value '%s' cannot be passed from one context to another. " +
-                            "The current context is 0x%x and the argument value originates from context 0x%x.",
-                            PolyglotValueDispatch.getValueInfo((PolyglotContextImpl) null, value), hashCode(), valueContext.hashCode()), nodeContext);
-        }
+    @TruffleBoundary
+    private RuntimeException failValueSharing() {
+        throw new ValueMigrationException("A value was tried to be migrated from one context to a different context. " +
+                        "Value migration for the current context was disabled and is therefore disallowed.", engine.getUncachedLocation());
     }
 
-    Object migrateHostWrapper(Node parentNode, PolyglotWrapper wrapper) {
+    Object migrateValue(Object value, PolyglotContextImpl valueContext) {
+        if (!config.allowValueSharing) {
+            throw failValueSharing();
+        }
+        Object result = engine.host.migrateValue(this, value, valueContext);
+        if (result != null) {
+            // host made sure migration is fine
+            return result;
+        }
+        // guaranteed by migrateValue
+        assert value instanceof TruffleObject;
+        if (value instanceof OtherContextGuestObject) {
+            OtherContextGuestObject otherValue = (OtherContextGuestObject) value;
+            if (otherValue.receiverContext == this && otherValue.delegateContext == valueContext) {
+                // reuse wrapper it is already wrapped
+                return otherValue;
+            } else if (otherValue.receiverContext == valueContext && otherValue.delegateContext == this) {
+                // unpack foreign value it belongs to that context
+                return otherValue.delegate;
+            } else {
+                return new OtherContextGuestObject(this, otherValue.delegate, valueContext);
+            }
+        }
+        assert value instanceof TruffleObject;
+        return new OtherContextGuestObject(this, value, valueContext);
+    }
+
+    Object migrateHostWrapper(PolyglotWrapper wrapper) {
         Object wrapped = wrapper.getGuestObject();
         PolyglotContextImpl valueContext = wrapper.getContext();
         if (valueContext != this) {
             // migrate wrapped value to the context
-            wrapped = migrateValue(parentNode, wrapped, valueContext);
+            wrapped = migrateValue(wrapped, valueContext);
         }
         return wrapped;
     }
@@ -1598,7 +1611,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
             } else {
                 targetLanguageContext = languageContext;
             }
-            return targetLanguageContext.asValue(toGuestValue(null, hostValue));
+            return targetLanguageContext.asValue(toGuestValue(hostValue));
         } catch (Throwable e) {
             throw PolyglotImpl.guestToHostException(this.getHostContext(), e, true);
         } finally {
@@ -1606,18 +1619,18 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
         }
     }
 
-    Object toGuestValue(Node location, Object hostValue) {
+    Object toGuestValue(Object hostValue) {
         if (hostValue instanceof Value) {
             Value receiverValue = (Value) hostValue;
             PolyglotLanguageContext languageContext = (PolyglotLanguageContext) getAPIAccess().getContext(receiverValue);
             PolyglotContextImpl valueContext = languageContext != null ? languageContext.context : null;
             Object valueReceiver = getAPIAccess().getReceiver(receiverValue);
             if (valueContext != this) {
-                valueReceiver = this.migrateValue(location, valueReceiver, valueContext);
+                valueReceiver = this.migrateValue(valueReceiver, valueContext);
             }
             return valueReceiver;
         } else if (PolyglotWrapper.isInstance(hostValue)) {
-            return migrateHostWrapper(location, PolyglotWrapper.asInstance(hostValue));
+            return migrateHostWrapper(PolyglotWrapper.asInstance(hostValue));
         } else {
             return engine.host.toGuestValue(getHostContextImpl(), hostValue);
         }
@@ -1994,7 +2007,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
             PolyglotContextImpl c = this;
             while (!threadInfo.explicitContextStack.isEmpty()) {
                 PolyglotContextImpl prev = threadInfo.explicitContextStack.removeLast();
-                engine.leave(prev, c, true);
+                engine.leave(prev, c);
                 c = prev;
             }
         }
@@ -2015,7 +2028,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
             assert !state.isClosed();
             PolyglotContextImpl prev;
             try {
-                prev = engine.enter(this, false, engine.getUncachedLocation(), !cancelOperation, cancelOperation);
+                prev = this.enterThreadChanged(false, false, !cancelOperation, cancelOperation);
             } catch (Throwable t) {
                 synchronized (this) {
                     restoreFromClosingState(cancelOperation);
@@ -2048,7 +2061,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
                      * synchronized.
                      */
                     assert !success || childContextsClosed() : "Polyglot context close marked as successful, but there are unclosed child contexts.";
-                    engine.leave(prev, this, false);
+                    this.leaveThreadChanged(prev, false, true);
                     if (success) {
                         remainingThreads = threads.keySet().toArray(new Thread[0]);
                     }
@@ -2578,7 +2591,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
         if (!newConfig.logLevels.isEmpty()) {
             EngineAccessor.LANGUAGE.configureLoggers(this, newConfig.logLevels, getAllLoggers());
         }
-        final PolyglotContextImpl prev = engine.enter(this, true, engine.getUncachedLocation(), true, false);
+        final PolyglotContextImpl prev = engine.enter(this);
         try {
             for (int i = 0; i < this.contexts.length; i++) {
                 final PolyglotLanguageContext context = this.contexts[i];
@@ -2590,7 +2603,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
                 }
             }
         } finally {
-            engine.leave(prev, this, true);
+            engine.leave(prev, this);
         }
         return true;
     }
@@ -2668,7 +2681,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
                         allowedLanguages,
                         Collections.emptyMap(),
                         fs, internalFs, engine.logHandler, false, null,
-                        EnvironmentAccess.INHERIT, null, null, null, null, null);
+                        EnvironmentAccess.INHERIT, null, null, null, null, null, true);
 
         final PolyglotContextImpl context = new PolyglotContextImpl(engine, config);
         synchronized (engine.lock) {
@@ -2701,7 +2714,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
             if (!languagesToPreinitialize.isEmpty()) {
                 context.inContextPreInitialization = true;
                 try {
-                    PolyglotContextImpl prev = context.engine.enter(context, true, context.engine.getUncachedLocation(), true, false);
+                    PolyglotContextImpl prev = context.engine.enter(context);
                     try {
                         for (String languageId : engine.getLanguages().keySet()) {
                             if (languagesToPreinitialize.contains(languageId)) {
@@ -2769,8 +2782,7 @@ final class PolyglotContextImpl implements com.oracle.truffle.polyglot.PolyglotI
                 LANGUAGE.disposeThread(languageContext.env, thread);
             }
         }
-
-        engine.leave(prev, this, true);
+        engine.leave(prev, this);
         assert !info.isActive();
 
         if (cachedThreadInfo.getThread() == thread) {